Add LogRunModel method and request class for model logging
A new `LogRunModel` method was added to the `IMLflowHttpClient`
interface in the `MLflowClient.Interfaces` namespace. This method
accepts a `LogRunModelRequest` object and returns a `Task`.

The `MLflowHttpClient` class in the `MLflowClient` namespace was
updated to implement the `LogRunModel` method, which sends a
request to the `MLflowApi.RunsLogModel` endpoint using the
`_apiClient.Post` function.

A new `LogRunModelRequest` class was introduced in the
`MLflowClient.Http.Dto.Runs.Request` namespace. This class
represents the request payload for the `LogRunModel` method,
including `RunId` and `ModelJson` properties. JSON serialization
is handled using the `Newtonsoft.Json` library.